refactor code

+++ src/main/java/com/codewithisa/filmservice/service/FilmServiceImpl.java
@@ -4,12 +4,13 @@ import com.codewithisa.filmservice.VO.ResponseTemplateFSSVO;
 import com.codewithisa.filmservice.VO.ResponseTemplateFSVO;
 import com.codewithisa.filmservice.VO.Schedules;
 import com.codewithisa.filmservice.VO.Seats;
-import com.codewithisa.filmservice.entity.Films;
+import com.codewithisa.filmservice.entity.Film;
 import com.codewithisa.filmservice.repository.FilmRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
@@ -19,27 +20,44 @@ import java.util.List;
 @Slf4j
 @Service
 public class FilmServiceImpl implements FilmService{
+
     @Autowired
     FilmRepository filmRepository;
+
     @Autowired
     RestTemplate restTemplate;
+
     @Override
-    public Films saveFilm(Films film) {
-        log.info("Inside saveFilm of FilmServiceImpl");
+    public Film saveFilm(Film film) throws Exception{
+        Boolean filmNameExists = existsByFilmName(film.getFilmName());
+        if(filmNameExists){
+            log.error("film name is already in the database");
+            throw new Exception("film name is already in the database");
+        }
         return filmRepository.save(film);
     }
 
     @Override
-    public Films findFilmByFilmCode(Long filmCode) {
-        log.info("Inside getFilmByFilmCode of FilmServiceImpl");
+    public Film findFilmByFilmCode(Long filmCode) throws Exception{
+        Boolean filmCodeExists = existsByFilmCode(filmCode);
+        if(!filmCodeExists){
+            log.error("film code is not exist");
+            throw new Exception("film code is not exist");
+        }
         return filmRepository.findFilmByFilmCode(filmCode);
     }
 
     @Override
-    public ResponseTemplateFSVO findFilmWithSchedules(Long filmCode) {
-        log.info("Inside findFilmWithSchedules of FilmServiceImpl");
+    public ResponseTemplateFSVO findFilmWithSchedules(Long filmCode) throws Exception{
+        Boolean filmCodeExists = existsByFilmCode(filmCode);
+
+        if(!filmCodeExists){
+            log.error("film code is not exist");
+            throw new Exception("film code is not exist");
+        }
+
         ResponseTemplateFSVO vo = new ResponseTemplateFSVO();
-        Films film = filmRepository.findFilmByFilmCode(filmCode);
+        Film film = filmRepository.findFilmByFilmCode(filmCode);
 
         ResponseEntity<List<Schedules>> schedulesList = restTemplate.exchange(
                 "https://schedule-service-production-12cd.up.railway.app/schedules/find-schedules-by-film-code/" + filmCode,
@@ -54,10 +72,16 @@ public class FilmServiceImpl implements FilmService{
     }
 
     @Override
-    public ResponseTemplateFSSVO findFilmWithScheduleAndSeats(Long filmCode, Long scheduleId) {
-        log.info("Inside findFilmWithSchedulesAndSeats of FilmServiceImpl");
+    public ResponseTemplateFSSVO findFilmWithScheduleAndSeats(Long filmCode, Long scheduleId) throws Exception{
+
+        Boolean filmCodeExists = existsByFilmCode(filmCode);
+        if(!filmCodeExists){
+            log.error("film code is not exist");
+            throw new Exception("film code is not exist");
+        }
+
         ResponseTemplateFSSVO vo = new ResponseTemplateFSSVO();
-        Films film = filmRepository.findFilmByFilmCode(filmCode);
+        Film film = filmRepository.findFilmByFilmCode(filmCode);
 
         Schedules schedule = restTemplate.getForObject(
                 "https://schedule-service-production-12cd.up.railway.app/schedules/find-schedule-by-schedule-id/" + scheduleId,
@@ -77,30 +101,56 @@ public class FilmServiceImpl implements FilmService{
     }
 
     @Override
-    public Films updateFilmName(Films film, Long filmCode) {
-        log.info("Inside updateFilmName of FilmServiceImpl");
-        Films existingFilm = filmRepository.findFilmByFilmCode(filmCode);
+    public Film updateFilmName(Film film, Long filmCode) throws Exception{
+        Boolean filmCodeExists = existsByFilmCode(filmCode);
+        if(!filmCodeExists){
+            log.error("film code is not exist");
+            throw new Exception("film code is not exist");
+        }
+
+        Boolean filmNameExists = existsByFilmName(film.getFilmName());
+        if(filmNameExists){
+            log.error("film name is already in the database");
+            throw new Exception("film name is already in the database");
+        }
+
+        Film existingFilm = filmRepository.findFilmByFilmCode(filmCode);
+
         existingFilm.setFilmName(film.getFilmName());
+
         filmRepository.save(existingFilm);
+
         log.info("film name is successfully updated");
+
         return existingFilm;
     }
 
     @Override
-    public void deleteFilm(Long filmCode) {
-        log.info("Inside deleteFilm of FilmServiceImpl");
+    public void deleteFilm(Long filmCode) throws Exception{
+        Boolean filmCodeExists = existsByFilmCode(filmCode);
+
+        if(!filmCodeExists){
+            log.error("film code is not exist");
+            throw new Exception("film code is not exist");
+        }
+
         filmRepository.deleteFilmByFilmCode(filmCode);
+        log.info("Film successfully deleted");
     }
 
     @Override
-    public List<Films> findFilmsYangSedangTayang() {
-        log.info("Inside findFilmsYangSedangTayang of FilmServiceImpl");
+    public List<Film> findFilmsYangSedangTayang() {
         return filmRepository.findFilmsYangSedangTayang();
     }
 
     @Override
-    public Films findFilmByFilmName(String filmName) {
-        log.info("Inside findFilmByFilmName of FilmServiceImpl");
+    public Film findFilmByFilmName(String filmName) throws Exception{
+        Boolean filmNameExists = existsByFilmName(filmName);
+
+        if(!filmNameExists){
+            log.error("film name is not exist in the database");
+            throw new Exception("film name is not exist in the database");
+        }
         return filmRepository.findFilmByFilmName(filmName);
     }
